Adds cvar sys_basedir and sys_homedir to get q2pro directories "basedir" and "homedir" to check map existance.
Also adds TestMap() function to g_main.c, to facilitate the test if a
file exists, starting in the basedir, then trying the homedir and
finally try the current directory.

And finally, in EndDMLevel() of g_main.c, adds code to change to the
current map if the file from the sv_maplist can not be found.